Get initial list of MineZones on each turn
A MineZone is created for each cleared cell with surrounding mines,
after each server response.

Added a function to compare two lists of MineZones, and try to
determine if the new one is better than the old one. Hopefully this
can be used in conjunction with a series of operations, e.g.
    compare(zone_sets([a, b], [a-b, b-a, a&b])

Server no longer allows attempts to clear already-cleared cells.