Started internal python server
Broken at the moment; AI seems to "give up" a lot of the time... This
may be because the server doesn't (yet) clear zeroes, and reactive_ai's
handling of this could be broken. Need investigation.

Also, all python files are currently in /client/py; internal_server.py
and server_json_wrapper.py should probably be in /server/ somewhere.