Multiprocessing games
Fixed the problem with games not winning/finishing: the cells_rem
count was incorrect on the internal python server, so games were never
"won".

Switched from JSON server to internal; it's much faster for many small
games, and still about twice as fast for a small number of large games.

Using multiprocessing. Plotting is only 10-20% faster on 2 cores, but
with more cores this should be much faster. Lost the progress counter,
for now.