"Balanced" empty-averaging client
Combination of the two EmptyAvg clients; switched between their
logics based on the ratio of mines to empty unknowns. Gets as good
results, for a 7x7 grid at least.

Not sure what governs the boundary ratio, but it seems to hold for a
number of different 2D and 3D game sizes I've tried.