Move guessing behaviour to subclass
Like in reactive_ai.py, Client once again doesn't guess - so I can
keep working on the no-guess logic. Guessing behaviour is in
subclass GuessClient.

ClientName is now a non-static, since statics can't be referenced
from instances, making it unfeasible for a subclass to provide it
own name. New games currently don't provide a client name to the
server until the first turn, as a result.